Add files via upload
Cost of Carry Model for researching the relationship between spot price and future price, using crude oil as an example, with the calculation of implied convenience yield with the observable future price. This model did not account for storage cost - which can be determined and be useful for long-term investment.